# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: MatchForge MVP API
  version: 0.3.0
  description: Read-only API for fixtures, immutable MVP forecasts, context, performance, and external prediction evidence.
servers:
  - url: http://127.0.0.1:8080
paths:
  /v1/competitions:
    get:
      operationId: listCompetitions
      parameters:
        - { name: continent, in: query, schema: { type: string } }
        - { name: country, in: query, schema: { type: string } }
        - { name: division, in: query, schema: { type: integer, minimum: 1 } }
        - { name: forecast_available, in: query, schema: { type: boolean } }
      responses:
        '200': { description: Discovered competitions and field availability, content: { application/json: { schema: { $ref: '#/components/schemas/CompetitionPage' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '503': { $ref: '#/components/responses/Unavailable' }
  /v1/fixtures:
    get:
      operationId: listFixtures
      parameters:
        - { name: date, in: query, required: true, schema: { type: string, format: date } }
        - { name: competition, in: query, schema: { type: string, format: uuid } }
        - { name: country, in: query, schema: { type: string } }
        - { name: continent, in: query, schema: { type: string } }
        - { name: forecast_availability, in: query, schema: { $ref: '#/components/schemas/ForecastAvailability' } }
      responses:
        '200': { description: Fixtures grouped by competition, content: { application/json: { schema: { $ref: '#/components/schemas/FixturePage' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '503': { $ref: '#/components/responses/Unavailable' }
  /v1/fixtures/{fixture_id}/context:
    get:
      operationId: getFixtureContext
      parameters: [ { $ref: '#/components/parameters/FixtureId' } ]
      responses:
        '200': { description: Display-only recent form, H2H, standings, and availability, content: { application/json: { schema: { type: object } } } }
        '503': { $ref: '#/components/responses/Unavailable' }
  /v1/fixtures/{fixture_id}/forecasts/{forecast_id}:
    get:
      operationId: getForecast
      parameters:
        - { $ref: '#/components/parameters/FixtureId' }
        - { name: forecast_id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Immutable pre-kickoff MVP forecast, content: { application/json: { schema: { $ref: '#/components/schemas/Forecast' } } } }
        '503': { $ref: '#/components/responses/Unavailable' }
  /v1/competitions/{competition_id}/standings:
    get:
      operationId: getStandings
      parameters: [ { $ref: '#/components/parameters/CompetitionId' } ]
      responses:
        '200': { description: Provider or completed-results fallback standings, content: { application/json: { schema: { type: object } } } }
        '503': { $ref: '#/components/responses/Unavailable' }
  /v1/competitions/{competition_id}/performance:
    get:
      operationId: getCompetitionPerformance
      parameters: [ { $ref: '#/components/parameters/CompetitionId' } ]
      responses:
        '200': { description: Chronological metrics and prior-results baseline comparison, content: { application/json: { schema: { $ref: '#/components/schemas/Performance' } } } }
        '503': { $ref: '#/components/responses/Unavailable' }
  /v1/external-predictions:
    get:
      operationId: listExternalPredictions
      parameters:
        - { name: date, in: query, schema: { type: string, format: date } }
        - { name: source, in: query, schema: { type: string } }
        - { name: competition, in: query, schema: { type: string } }
        - { name: market, in: query, schema: { type: string } }
        - { name: agreement, in: query, schema: { type: string, enum: [AGREES, WEAK_SUPPORT, DISAGREES, UNABLE_TO_EVALUATE] } }
        - { name: fixture_id, in: query, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Stored source selections and independent comparison, content: { application/json: { schema: { type: object } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '503': { $ref: '#/components/responses/Unavailable' }
  /v1/fixtures/{fixture_id}/external-predictions:
    get:
      operationId: listFixtureExternalPredictions
      parameters: [ { $ref: '#/components/parameters/FixtureId' } ]
      responses:
        '200': { description: Stored selections for one mapped fixture, content: { application/json: { schema: { type: object } } } }
        '503': { $ref: '#/components/responses/Unavailable' }
  /v1/external-prediction-sources:
    get:
      operationId: listExternalPredictionSources
      responses:
        '200': { description: Access audit, adapter state, and settled tracking metrics, content: { application/json: { schema: { type: object } } } }
        '503': { $ref: '#/components/responses/Unavailable' }
components:
  parameters:
    FixtureId: { name: fixture_id, in: path, required: true, schema: { type: string, format: uuid } }
    CompetitionId: { name: competition_id, in: path, required: true, schema: { type: string, format: uuid } }
  responses:
    BadRequest: { description: Invalid request, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    Unavailable: { description: Product data unavailable, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  schemas:
    Error: { type: object, required: [error], properties: { error: { type: string } } }
    ForecastAvailability: { type: string, enum: [FORECAST_AVAILABLE, NOT_ENOUGH_HISTORY, SOURCE_DATA_INCOMPLETE, TEMPORARILY_UNAVAILABLE] }
    Availability:
      type: object
      required: [fixtures, results, standings, h2h, forecast, xg, team_stats]
      properties:
        fixtures: { type: boolean }
        results: { type: boolean }
        standings: { type: boolean }
        h2h: { type: boolean }
        forecast: { type: boolean }
        xg: { type: boolean }
        team_stats: { type: boolean }
    Competition:
      type: object
      required: [id, name, country, continent, type, season, availability, availability_status, sources]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        country: { type: string }
        continent: { type: string }
        division: { type: [integer, 'null'] }
        type: { type: string, enum: [LEAGUE, CUP] }
        season: { type: string }
        availability: { $ref: '#/components/schemas/Availability' }
        availability_status: { $ref: '#/components/schemas/ForecastAvailability' }
        sources: { type: object, additionalProperties: { type: array, items: { type: string } } }
    CompetitionPage:
      type: object
      required: [competitions]
      properties: { competitions: { type: array, items: { $ref: '#/components/schemas/Competition' } } }
    Team:
      type: object
      required: [id, name, crest_url]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        crest_url: { type: [string, 'null'], format: uri }
    ForecastSummary:
      type: object
      required: [id, expected_home_goals, expected_away_goals, probabilities]
      properties:
        id: { type: string, format: uuid }
        expected_home_goals: { type: number, exclusiveMinimum: 0 }
        expected_away_goals: { type: number, exclusiveMinimum: 0 }
        probabilities: { type: object, additionalProperties: { type: number, minimum: 0, maximum: 1 } }
    Fixture:
      type: object
      required: [id, kickoff_at, status, home, away, home_score, away_score, venue, round, forecast_availability, forecast]
      properties:
        id: { type: string, format: uuid }
        kickoff_at: { type: string, format: date-time }
        status: { type: string, enum: [SCHEDULED, LIVE, FINISHED, POSTPONED, CANCELLED, ABANDONED] }
        home: { $ref: '#/components/schemas/Team' }
        away: { $ref: '#/components/schemas/Team' }
        home_score: { type: [integer, 'null'], minimum: 0 }
        away_score: { type: [integer, 'null'], minimum: 0 }
        venue: { type: [string, 'null'] }
        round: { type: [string, 'null'] }
        forecast_availability: { $ref: '#/components/schemas/ForecastAvailability' }
        forecast: { anyOf: [ { $ref: '#/components/schemas/ForecastSummary' }, { type: 'null' } ] }
    FixtureGroup:
      type: object
      required: [competition, fixtures]
      properties:
        competition: { $ref: '#/components/schemas/Competition' }
        fixtures: { type: array, items: { $ref: '#/components/schemas/Fixture' } }
    FixturePage:
      type: object
      required: [date, groups]
      properties:
        date: { type: string, format: date }
        groups: { type: array, items: { $ref: '#/components/schemas/FixtureGroup' } }
    Forecast:
      allOf:
        - { $ref: '#/components/schemas/ForecastSummary' }
        - type: object
          required: [fixture_id, model_label, model_algorithm_version, created_at, football_cutoff, knowledge_cutoff, knowledge_mode, publication_mode, score_matrix]
          properties:
            fixture_id: { type: string, format: uuid }
            model_label: { type: string, const: MVP_FORECAST }
            model_algorithm_version: { type: string }
            created_at: { type: string, format: date-time }
            football_cutoff: { type: string, format: date-time }
            knowledge_cutoff: { type: string, format: date-time }
            knowledge_mode: { type: string, const: bitemporal }
            publication_mode: { type: string, const: MVP_OWNER_AUTHORIZED }
            score_matrix: { type: array, items: { type: object } }
    Performance:
      type: object
      required: [competition_id, rating, forecasts]
      properties:
        competition_id: { type: string, format: uuid }
        rating: { type: string, enum: [UNRATED, WATCH, GOOD, STRONG] }
        forecasts: { type: integer, minimum: 0 }
        brier: { type: [number, 'null'] }
        log_loss: { type: [number, 'null'] }
        baseline_brier: { type: [number, 'null'] }
        baseline_log_loss: { type: [number, 'null'] }