# go/api/internal/app/app.go
package app

import (
	"context"
	"encoding/json"
	"errors"
	"log/slog"
	"net/http"
)

// App owns the operational HTTP surface and lifecycle.
type App struct {
	config    Config
	logger    *slog.Logger
	readiness ReadinessChecker
	product   ProductStore
	handler   http.Handler
}

// New constructs the operational API without starting network listeners.
func New(config Config, logger *slog.Logger, readiness ReadinessChecker, stores ...ProductStore) *App {
	var product ProductStore = unavailableProductStore{}
	if len(stores) > 0 && stores[0] != nil {
		product = stores[0]
	}
	application := &App{config: config, logger: logger, readiness: readiness, product: product}
	mux := http.NewServeMux()
	mux.HandleFunc("GET /healthz", application.health)
	mux.HandleFunc("GET /readyz", application.ready)
	mux.HandleFunc("GET /version", application.version)
	mux.HandleFunc("GET /v1/competitions", application.listCompetitions)
	mux.HandleFunc("GET /v1/fixtures", application.listFixtures)
	mux.HandleFunc("GET /v1/fixtures/{fixture_id}/context", application.fixtureContext)
	mux.HandleFunc("GET /v1/fixtures/{fixture_id}/forecasts/{forecast_id}", application.forecast)
	mux.HandleFunc("GET /v1/competitions/{competition_id}/standings", application.standings)
	mux.HandleFunc("GET /v1/competitions/{competition_id}/performance", application.performance)
	mux.HandleFunc("GET /v1/external-predictions", application.externalPredictions)
	mux.HandleFunc("GET /v1/fixtures/{fixture_id}/external-predictions", application.fixtureExternalPredictions)
	mux.HandleFunc("GET /v1/external-prediction-sources", application.externalPredictionSources)
	application.handler = application.cors(mux)
	return application
}

// Handler exposes the HTTP contract for tests and servers.
func (application *App) Handler() http.Handler {
	return application.handler
}

// Run serves until the context is cancelled, then performs bounded graceful shutdown.
func (application *App) Run(ctx context.Context) error {
	server := &http.Server{
		Addr:              application.config.Address,
		Handler:           application.handler,
		ReadHeaderTimeout: application.config.ReadHeaderTimeout,
	}
	errorsChannel := make(chan error, 1)
	go func() {
		application.logger.Info("api listening", "address", application.config.Address)
		errorsChannel <- server.ListenAndServe()
	}()

	select {
	case err := <-errorsChannel:
		if errors.Is(err, http.ErrServerClosed) {
			return nil
		}
		return err
	case <-ctx.Done():
		shutdownContext, cancel := context.WithTimeout(context.Background(), application.config.ShutdownTimeout)
		defer cancel()
		return server.Shutdown(shutdownContext)
	}
}

func (application *App) health(response http.ResponseWriter, _ *http.Request) {
	application.writeJSON(response, http.StatusOK, map[string]string{"status": "ok"})
}

func (application *App) ready(response http.ResponseWriter, request *http.Request) {
	if err := application.readiness.Ready(request.Context()); err != nil {
		application.logger.Warn("readiness failed", "error", err)
		application.writeJSON(response, http.StatusServiceUnavailable, map[string]string{"status": "unavailable"})
		return
	}
	application.writeJSON(response, http.StatusOK, map[string]string{"status": "ready"})
}

func (application *App) version(response http.ResponseWriter, _ *http.Request) {
	application.writeJSON(response, http.StatusOK, map[string]string{"version": application.config.Version})
}

func (application *App) writeJSON(response http.ResponseWriter, status int, value any) {
	response.Header().Set("Content-Type", "application/json")
	response.WriteHeader(status)
	if err := json.NewEncoder(response).Encode(value); err != nil {
		application.logger.Error("write JSON response", "error", err)
	}
}

func (application *App) cors(next http.Handler) http.Handler {
	return http.HandlerFunc(func(response http.ResponseWriter, request *http.Request) {
		if request.Header.Get("Origin") == application.config.AllowedOrigin {
			response.Header().Set("Access-Control-Allow-Origin", application.config.AllowedOrigin)
			response.Header().Set("Vary", "Origin")
		}
		next.ServeHTTP(response, request)
	})
}